+++ ui/graphicsviewzoom.h
@@ -21,7 +21,8 @@
 
 #include <QObject>
 #include <QGraphicsView>
-#include "program/globals.h"
+
+enum ZoomSource {MOUSE_WHEEL, SPIN_BOX, KEYBOARD, GESTURE};
 
 // This class is a modified version of:
 // http://stackoverflow.com/questions/19113532/qgraphicsview-zooming-in-and-out-under-mouse-position-using-mouse-wheel
